Anagram Checker
[Anagram Checker] Write a program to check whether two strings are anagrams
or not. Each String is assumed to consist of alphabets only. Two words are said to
be anagrams of each other if the letters from one word can be rearranged to form
the other word. For example "abcc" and "cabc" are anagram strings but “abcc” and
“cabb” are not anagrams.

Sample Input: s1 = “”mississipi”, s2 = “ssissipimi”
Sample output: s1 and s2 are anagrams.